Added test device info
Google is quite strict with invalid ad requests and can block an account
if detects impressions or clicks that are supposed not to be. Plugin now
adds the emulator flag to be excluded from Ad request and, there is also
a new flag, called isTestDevice. When this flag is TRUE, the AdMob's
device id is calculated and set into the AdRequest to be able to test
the application without any concern with admob..